+++ src/components/index.ts
@@ -1,5 +1,4 @@
 export * from './FieldsEditor';
-export * from './FrameReducer';
 export * from './NullableInput';
 export * from './QueryEditor';
 export * from './ValuesEditor';