+++ src/constants/index.ts
@@ -1,3 +1,2 @@
 export * from './datasource';
 export * from './field';
-export * from './visualization';